backend/requirements.txt
langgraph==0.2.45
langchain-groq==0.2.1
langchain-core==0.3.21
tavily-python==0.5.0
python-dotenv==1.0.0
pydantic==2.10.3
colorama==0.4.6
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sse-starlette==2.1.3
